requirements.txt
pytest==7.4.4
pytest-bdd==7.1.2
playwright==1.41.2
pytest-html==4.1.1
python-dotenv==1.0.1

